🌱 split useStackDiscovery: extract types and filters
